third_party/blink/web_tests/http/tests/security/promise-access-control-allow-expected.txt

CONSOLE ERROR: Uncaught (in promise) 42
CONSOLE MESSAGE: PASS: received unhandledrejection event
CONSOLE MESSAGE: PASS: received expected reason
Tests that promise rejection events are received for cross origin CORS scripts
